services/anomaly_detector.py
from sklearn.ensemble import IsolationForest
import numpy as np
from typing import Dict

class EarningsAnomalyDetector:
    def __init__(self):
        # Initialize and potentially train the model here
        # In a real application, the model would be trained on historical data and loaded.
        # For this example, we'll initialize a simple model.
        self.model = IsolationForest(contamination='auto', random_state=42)
        # Placeholder for training - ideally load a pre-trained model
        # self.model.fit(historical_data)

    def check(self, claim_data: Dict) -> bool:
        """Check if the claim data is an anomaly"""
        # Features for anomaly detection
        features = [
            claim_data.get('earnings', 0.0),
            claim_data.get('employment_months', 0),
            # Using a simple proxy for employer legitimacy/size - could be enhanced
            len(claim_data.get('employer', ''))
        ]
        
        # Predict anomaly (-1 for anomaly, 1 for inlier)
        prediction = self.model.predict([features])[0]
        
        return prediction == -1

    # Placeholder for retraining logic
    # def retrain(self, historical_data):
    #     self.model.fit(historical_data) 