requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.6.1
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.1
streamlit==1.31.1
numpy==1.26.4 